updated pkix/pkix.go retaining changes in pkix_modified.go
